feat(botcoin): Update to new strategy
Configure a list of buy orders and a profit percentage.
As soon as the first buy order is filled, a sell order will be created
for the profit percentage above the price of the first buy order.
When a second buy order is filled, the sell order will be updated to be
for the profit percentage above the average buy price.